feat(affiliate code updates): Affiliate code updates
