+++ src/index.ts
@@ -3,4 +3,6 @@ import Fuul from './core';
 export type { Conversion } from './types/api';
 export type { EventArgs, FuulSettings, UserMetadata } from './types/sdk';
 
+export * from './affiliates/errors';
+
 export { Fuul };
